+++ src/events/selectMenuUsed.ts
@@ -6,11 +6,7 @@ export const event: Event = {
 	run: async (client: Client, interaction: SelectMenuInteraction) => {
 		client.storeAction("selectMenu", interaction);
 
-		let selectMenu: SelectMenu;
-
-		if (interaction.customId.startsWith("textureSelect_")) selectMenu = client.menus.get("textureSelect");
-		else selectMenu = client.menus.get(interaction.customId);
-
+		const selectMenu: SelectMenu = client.menus.get(interaction.customId.split("_")[0])
 		if (selectMenu) return selectMenu.execute(client, interaction);
 	},
 };